+++ src/tk/PWinObj.hh
@@ -202,7 +202,8 @@ public:
 	/** Set focusable flag. */
 	virtual void setFocusable(bool focusable) { _focusable = focusable; }
 
-	virtual void giveInputFocus(void);
+	virtual void giveInputFocus();
+	virtual void warpPointer();
 	virtual void reparent(PWinObj *parent, int x, int y);
 
 	virtual bool getSizeRequest(Geometry &request);